[level 1] Title: 폰켓몬, Time: 0.73 ms, Memory: 10.1 MB -BaekjoonHub

+++ 프로그래머스/1/1845. 폰켓몬/폰켓몬.py
@@ -6,8 +6,8 @@ def solution(nums):
     # 2. 타입 종류 확인 O(1)
     type_count = len(type_set)
     # 3. 최대 선택 갯수 확인 O(1)
-    # 만약 nums 길이가 짝수가 아니었으면 내림이 되니까 그런 케이스일 경우 주의!
-    max_pick = len(nums)/2
+    # 만약 nums 길이가 짝수가 아닌 경우 /2했을 때 어떻게 처리하는지 디렉션이 있을 경우 주의!
+    max_pick =int(len(nums)/2)
     
     # 4. 가장 많은 종류 선택 로직 O(1)
     return min(type_count, max_pick)
